refactor: cli to ts
