+++ packages/ordi/cli/webpack/client/webpack.prod.ts
@@ -1,0 +1,14 @@
+import { type Configuration } from "webpack";
+import { mergeWithCustomize, customizeObject } from "webpack-merge";
+import common from "./webpack.config";
+
+export default mergeWithCustomize<Configuration>({
+  customizeObject: customizeObject({
+    output: "append",
+  }),
+})(common, {
+  mode: "production",
+  output: {
+    filename: "[name].[contenthash].js",
+  },
+});
