+++ packages/ordi/cli/webpack/webpack.shared.ts
@@ -1,13 +1,10 @@
-const webpack = require("webpack");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+import webpack from "webpack";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+import resolver from "./resolve";
+import defaultProcessEnv from "./plugins/DefinePlugin";
+import ifDev from "../../src/utils/ifDev";
 
-const { default: ifDev } = require("../../src/utils/ifDev");
-
-const resolver = require("./resolve");
-const defaultProcessEnv = require("./plugins/DefinePlugin").default;
-
-
-module.exports = {
+const shared = {
   resolve: {
     extensions: [".ts", ".tsx", ".js", ".jsx"],
     symlinks: true,
@@ -37,14 +34,17 @@ module.exports = {
           options: {
             presets: [
               "@babel/preset-typescript",
-              ["@babel/preset-react", { runtime: "automatic", useBuiltIns: true }],
+              [
+                "@babel/preset-react",
+                { runtime: "automatic", useBuiltIns: true },
+              ],
               [
                 "@babel/preset-env",
                 {
                   useBuiltIns: "entry",
                   targets: "> 0.25%, not dead",
                   corejs: 3,
-                  modules: false
+                  modules: false,
                 },
               ],
             ],
@@ -86,3 +86,5 @@ module.exports = {
     ],
   },
 };
+
+export default shared;
