+++ packages/ordi/package.json
@@ -55,6 +55,7 @@
     "@types/jest": "^27.5.2",
     "@types/loadable__component": "^5.13.9",
     "@types/loadable__server": "^5.12.11",
+    "@types/loadable__webpack-plugin": "^5.7.6",
     "@types/lru-cache": "^5.1.1",
     "@types/node": "^18.19.38",
     "@types/react": "^17.0.80",
